refactor(transaction): split the logo routes out of the catalogue resource
detekt's TooManyFunctions fired AT the threshold (11 of 11) when ingest added a
route, and the honest fix is the split it was pointing at rather than a suppress:
the catalogue answers who was paid, the logo routes answer what their mark looks
like, and one class holding both had passed the point where a reader keeps it in
their head. Same shape as MerchantLocationResource.

Also from that run, and each a real finding rather than a style opinion: the
public-address check's five-way condition is now a named predicate (what it
excludes is worth a name and a comment); every RefusedException thrown from a
catch carries its cause, so a refusal can still be traced to the IO error behind
it; and the two functions whose separate throws are separate operator-actionable
refusals say so with a reason instead of being collapsed.

(cherry picked from commit 3d0139183ca1a5b630b9fc523ad6bbc2c343bc43)
