+++ openbank-transaction-service/src/test/kotlin/com/openbank/transaction/infrastructure/rest/MerchantCatalogResourceTest.kt
@@ -44,6 +44,7 @@ class MerchantCatalogResourceTest {
     private lateinit var locations: MerchantLocationRepository
     private lateinit var locationResource: MerchantLocationResource
     private lateinit var fetcher: LogoFetcher
+    private lateinit var logoResource: MerchantLogoResource
     private lateinit var resource: MerchantCatalogResource
 
     @BeforeEach
@@ -55,7 +56,8 @@ class MerchantCatalogResourceTest {
         // Fetching OFF, which is the default everywhere: these tests are about the catalogue, and a
         // configured allowlist would put a real network call behind them.
         fetcher = LogoFetcher(java.util.Optional.empty())
-        resource = MerchantCatalogResource(catalog, transactions, logos, fetcher)
+        resource = MerchantCatalogResource(catalog, transactions)
+        logoResource = MerchantLogoResource(logos, fetcher)
         locationResource = MerchantLocationResource(locations)
         coEvery { catalog.upsert(any()) } returns true
         coEvery { catalog.findByKey(any()) } returns null
@@ -221,7 +223,7 @@ class MerchantCatalogResourceTest {
     fun `a raw descriptor resolves to the same logo as its normalised key`() {
         coEvery { logos.findByKey("ALZACZ") } returns logoEntity("ALZACZ", "a".repeat(64))
 
-        val response = runBlocking { resource.logo("ALZA.CZ A.S. PRAHA 4", 64, unconditionalRequest()) }
+        val response = runBlocking { logoResource.logo("ALZA.CZ A.S. PRAHA 4", 64, unconditionalRequest()) }
 
         assertThat(response.status).isEqualTo(200)
         coVerify { logos.findByKey("ALZACZ") }
@@ -235,7 +237,7 @@ class MerchantCatalogResourceTest {
      */
     @Test
     fun `an unsupported size is a 400, not a served variant`() {
-        val response = runBlocking { resource.logo("ALZACZ", 512, unconditionalRequest()) }
+        val response = runBlocking { logoResource.logo("ALZACZ", 512, unconditionalRequest()) }
 
         assertThat(response.status).isEqualTo(400)
     }
@@ -244,8 +246,8 @@ class MerchantCatalogResourceTest {
     fun `the two supported sizes serve the two stored variants`() {
         coEvery { logos.findByKey("ALZACZ") } returns logoEntity("ALZACZ", "b".repeat(64))
 
-        val small = runBlocking { resource.logo("ALZACZ", 64, unconditionalRequest()) }
-        val large = runBlocking { resource.logo("ALZACZ", 128, unconditionalRequest()) }
+        val small = runBlocking { logoResource.logo("ALZACZ", 64, unconditionalRequest()) }
+        val large = runBlocking { logoResource.logo("ALZACZ", 128, unconditionalRequest()) }
 
         assertThat(small.entity as ByteArray).hasSize(3)
         assertThat(large.entity as ByteArray).hasSize(4)
@@ -265,7 +267,7 @@ class MerchantCatalogResourceTest {
                 Response.notModified()
         }
 
-        val response = runBlocking { resource.logo("ALZACZ", 64, conditional) }
+        val response = runBlocking { logoResource.logo("ALZACZ", 64, conditional) }
 
         assertThat(response.status).isEqualTo(304)
         assertThat(response.entity).isNull()
@@ -277,7 +279,7 @@ class MerchantCatalogResourceTest {
     fun `a merchant with no stored logo is a 404, never a placeholder image`() {
         coEvery { logos.findByKey(any()) } returns null
 
-        val response = runBlocking { resource.logo("ALZACZ", 64, unconditionalRequest()) }
+        val response = runBlocking { logoResource.logo("ALZACZ", 64, unconditionalRequest()) }
 
         assertThat(response.status).isEqualTo(404)
     }
@@ -294,7 +296,7 @@ class MerchantCatalogResourceTest {
         coEvery { logos.upsert(capture(saved)) } returns true
 
         val response = runBlocking {
-            resource.putLogo("ALZA.CZ A.S.", null, "trademark", null, operator(), upload)
+            logoResource.putLogo("ALZA.CZ A.S.", null, "trademark", null, operator(), upload)
         }
 
         assertThat(response.status).isEqualTo(201)
@@ -313,7 +315,7 @@ class MerchantCatalogResourceTest {
      */
     @Test
     fun `a missing request body is a 400, not a 500`() {
-        val response = runBlocking { resource.putLogo("ALZACZ", null, null, null, operator(), null) }
+        val response = runBlocking { logoResource.putLogo("ALZACZ", null, null, null, operator(), null) }
 
         assertThat(response.status).isEqualTo(400)
     }
@@ -321,7 +323,7 @@ class MerchantCatalogResourceTest {
     @Test
     fun `an upload that is not a raster image is refused with the reason`() {
         val response = runBlocking {
-            resource.putLogo("ALZACZ", null, null, null, operator(), "<svg/>".toByteArray())
+            logoResource.putLogo("ALZACZ", null, null, null, operator(), "<svg/>".toByteArray())
         }
 
         assertThat(response.status).isEqualTo(400)
@@ -338,7 +340,7 @@ class MerchantCatalogResourceTest {
         coEvery { logos.upsert(any()) } returns null
 
         val response = runBlocking {
-            resource.putLogo("NEVERSEEN", null, null, null, operator(), pngBytes())
+            logoResource.putLogo("NEVERSEEN", null, null, null, operator(), pngBytes())
         }
 
         assertThat(response.status).isEqualTo(404)
@@ -348,7 +350,7 @@ class MerchantCatalogResourceTest {
     fun `deleting a logo that is not there is a 404`() {
         coEvery { logos.deleteByKey(any()) } returns false
 
-        val response = runBlocking { resource.deleteLogo("ALZACZ") }
+        val response = runBlocking { logoResource.deleteLogo("ALZACZ") }
 
         assertThat(response.status).isEqualTo(404)
     }
